refactor: utils file for helper functions
